Patch with prepend and check for existance
