+++ lib/inertia_rails.rb
@@ -1,6 +1,7 @@
 require 'inertia_rails/renderer'
 require 'inertia_rails/engine'
-require 'inertia_rails/debug_exceptions'
+
+require 'patches/debug_exceptions'
 
 ActionController::Renderers.add :inertia do |component, options|
   InertiaRails::Renderer.new(
